requirements.txt
python-barcode
Pillow
Flask
